Adjust to allow automatic payment process.
Add webhook and default way of loading configuration for sms and kopokopo

Adjust the ui for customer to easily input the number to launch stk push.

Adjust setting default config and ui for inserting the number.

Peform clean up and add sms functionality, v1 complete.

Adjust the way sms log is created as well as deal with connection issues.

Limit the amount to be sent a code.

Add workspace for the app.